requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14
pyautogui==0.9.54
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.5
customtkinter==5.2.0
Pillow==10.0.0
pytesseract==0.3.10
openai==1.3.0  # Used for Groq API (compatible with OpenAI SDK)
schedule==1.2.0
pyinstaller==6.2.0

